app/requirements.txt
django
cloudinary
django-cloudinary-storage
django-heroku
python-decouple
psycopg2-binary
django-environ
django-ckeditor
crispy-bootstrap5
django-crispy-forms
django-widget-tweaks
django-bootstrap-modal-forms
social-auth-app-django
pillow
gunicorn
djangorestframework
python-decouple
psycopg2-binary
django-rosetta
django-parler
django-hitcount
django-recaptcha
sendgrid
requests
django-tinymce
django-next-prev
#django-picklefield
django-json-widget
django-highlightjs

